+++ src/index.ts
@@ -7,3 +7,5 @@ export { UserErrorsException } from './exception/user-errors-exception';
 
 export type { FieldDefinition, DefinitionSchemaEntry, DefinitionSchema } from './types/definitions';
 export type { OnPopulateFunc } from './types/metaobject-repository';
+
+export { createFormState } from './utils/form';