+++ src/types/metaobject-repository.ts
@@ -5,7 +5,8 @@
  */
 
 import { FieldBuilder } from "raku-ql";
-import { CamelCase, CapabilityInputMap, DefaultMap, DefinitionByType, DefinitionSchema, FieldDefinition } from "./definitions";
+import { CamelCase, CapabilityInputMap, DefaultMap, DefinitionByType, DefinitionSchema, FieldDefinition, FromDefinitionWithSystemData, ValidPopulatePaths } from "./definitions";
+import { MetaobjectRepository } from "~/metaobject-repository";
 
 /**
  * --------------------------------------------------------------------------------------------